# blog/api/blog/v1/task.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "任务模块-接口文档",
    "description": "任务模块文档",
    "version": "1.0",
    "contact": {
      "name": "易起游saas 易镜项目",
      "url": "https://gitlab.stnts.com/shangyecelue/yiqiyoo",
      "email": "luoyuxiang2193@shengtian.com"
    }
  },
  "tags": [
    {
      "name": "TaskGrpcService"
    }
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/v1/task/day/list": {
      "get": {
        "summary": "每日任务详情",
        "description": "每日任务详情",
        "operationId": "TaskGrpcService_ListDayTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListDayTaskReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "每日任务模块"
        ],
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/v1/task/day/push": {
      "post": {
        "summary": "每日任务上报",
        "description": "每日任务上报",
        "operationId": "TaskGrpcService_PushDayTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PushDayTaskRequest"
            }
          }
        ],
        "tags": [
          "每日任务模块"
        ],
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/v1/task/day/recv": {
      "post": {
        "summary": "每日任务领取",
        "description": "每日任务领取",
        "operationId": "TaskGrpcService_RecvDayTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1RecvDayTaskRequest"
            }
          }
        ],
        "tags": [
          "每日任务模块"
        ],
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/v1/task/master/list": {
      "get": {
        "summary": "主线任务详情",
        "description": "主线任务详情",
        "operationId": "TaskGrpcService_ListMasterTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ListMasterTaskReply"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "主线任务模块"
        ],
        "security": [
          {
            "jwt": []
          }
        ]
      }
    },
    "/v1/task/master/push": {
      "post": {
        "summary": "主线任务上报",
        "description": "主线任务上报",
        "operationId": "TaskGrpcService_PushMasterTask",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1Empty"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1PushMasterTaskRequest"
            }
          }
        ],
        "tags": [
          "主线任务模块"
        ],
        "security": [
          {
            "jwt": []
          }
        ]
      }
    }
  },
  "definitions": {
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string",
          "description": "A URL/resource name that uniquely identifies the type of the serialized\nprotocol buffer message. This string must contain at least\none \"/\" character. The last segment of the URL's path must represent\nthe fully qualified name of the type (as in\n`path/google.protobuf.Duration`). The name should be in a canonical form\n(e.g., leading \".\" is not accepted).\n\nIn practice, teams usually precompile into the binary all types that they\nexpect it to use in the context of Any. However, for URLs which use the\nscheme `http`, `https`, or no scheme, one can optionally set up a type\nserver that maps type URLs to message definitions as follows:\n\n* If no scheme is provided, `https` is assumed.\n* An HTTP GET on the URL must yield a [google.protobuf.Type][]\n  value in binary format, or produce an error.\n* Applications are allowed to cache lookup results based on the\n  URL, or have them precompiled into a binary to avoid any\n  lookup. Therefore, binary compatibility needs to be preserved\n  on changes to types. (Use versioned type names to manage\n  breaking changes.)\n\nNote: this functionality is not currently available in the official\nprotobuf release, and it is not used for type URLs beginning with\ntype.googleapis.com.\n\nSchemes other than `http`, `https` (or the empty scheme) might be\nused with implementation specific semantics."
        }
      },
      "additionalProperties": {},
      "description": "`Any` contains an arbitrary serialized protocol buffer message along with a\nURL that describes the type of the serialized message.\n\nProtobuf library provides support to pack/unpack Any values in the form\nof utility functions or additional generated methods of the Any type.\n\nExample 1: Pack and unpack a message in C++.\n\n    Foo foo = ...;\n    Any any;\n    any.PackFrom(foo);\n    ...\n    if (any.UnpackTo(\u0026foo)) {\n      ...\n    }\n\nExample 2: Pack and unpack a message in Java.\n\n    Foo foo = ...;\n    Any any = Any.pack(foo);\n    ...\n    if (any.is(Foo.class)) {\n      foo = any.unpack(Foo.class);\n    }\n\nExample 3: Pack and unpack a message in Python.\n\n    foo = Foo(...)\n    any = Any()\n    any.Pack(foo)\n    ...\n    if any.Is(Foo.DESCRIPTOR):\n      any.Unpack(foo)\n      ...\n\nExample 4: Pack and unpack a message in Go\n\n     foo := \u0026pb.Foo{...}\n     any, err := anypb.New(foo)\n     if err != nil {\n       ...\n     }\n     ...\n     foo := \u0026pb.Foo{}\n     if err := any.UnmarshalTo(foo); err != nil {\n       ...\n     }\n\nThe pack methods provided by protobuf library will by default use\n'type.googleapis.com/full.type.name' as the type URL and the unpack\nmethods only use the fully qualified type name after the last '/'\nin the type URL, for example \"foo.bar.com/x/y.z\" will yield type\nname \"y.z\".\n\n\nJSON\n\nThe JSON representation of an `Any` value uses the regular\nrepresentation of the deserialized, embedded message, with an\nadditional field `@type` which contains the type URL. Example:\n\n    package google.profile;\n    message Person {\n      string first_name = 1;\n      string last_name = 2;\n    }\n\n    {\n      \"@type\": \"type.googleapis.com/google.profile.Person\",\n      \"firstName\": \u003cstring\u003e,\n      \"lastName\": \u003cstring\u003e\n    }\n\nIf the embedded message type is well-known and has a custom JSON\nrepresentation, that representation will be embedded adding a field\n`value` which holds the custom JSON in addition to the `@type`\nfield. Example (for message [google.protobuf.Duration][]):\n\n    {\n      \"@type\": \"type.googleapis.com/google.protobuf.Duration\",\n      \"value\": \"1.212s\"\n    }"
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "v1DayTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "active_id": {
          "type": "string"
        },
        "member_id": {
          "type": "string",
          "format": "int64"
        },
        "task_id": {
          "type": "string",
          "format": "int64"
        },
        "adddate": {
          "type": "string"
        },
        "task_name": {
          "type": "string"
        },
        "condition_num": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "rewards": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1Reward"
          }
        }
      },
      "title": "每日任务"
    },
    "v1DayTaskRecord": {
      "type": "object",
      "properties": {
        "day_task_id": {
          "type": "string",
          "format": "int64"
        },
        "condition": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "完成每日任务记录"
    },
    "v1Empty": {
      "type": "object",
      "title": "空对象"
    },
    "v1ListDayTaskReply": {
      "type": "object",
      "properties": {
        "active_val": {
          "type": "integer",
          "format": "int32"
        },
        "day_task_active_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1DayTask"
          }
        },
        "day_task_active_total_state": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1DayTaskRecord"
          }
        },
        "day_task_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1DayTask"
          }
        },
        "day_task_total_state": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1DayTaskRecord"
          }
        }
      }
    },
    "v1ListMasterTaskReply": {
      "type": "object",
      "properties": {
        "master_task_total_state": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1MasterTask"
          }
        },
        "master_task_list": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1MasterTaskRecord"
          }
        }
      }
    },
    "v1MasterTask": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "active_id": {
          "type": "string"
        },
        "task_id": {
          "type": "string",
          "format": "int64"
        },
        "task_name": {
          "type": "string"
        },
        "condition_num": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "integer",
          "format": "int32"
        },
        "config": {
          "type": "string"
        }
      },
      "title": "主线任务"
    },
    "v1MasterTaskRecord": {
      "type": "object",
      "properties": {
        "master_task_id": {
          "type": "string",
          "format": "int64"
        },
        "condition": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1PushDayTaskRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1PushMasterTaskRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1RecvDayTaskRequest": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "v1Reward": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string",
          "format": "int64"
        },
        "icon": {
          "type": "string"
        },
        "num": {
          "type": "integer",
          "format": "int32"
        }
      },
      "title": "奖励"
    }
  },
  "securityDefinitions": {
    "jwt": {
      "type": "apiKey",
      "description": "授权token",
      "name": "Authorization",
      "in": "header",
      "flow": "implicit"
    }
  }
}